fix(scanner): recycle playwright context + coerce v3 types
Two restart-triggered regressions in the post-weighted_score scanner path:

1. Playwright browser context degrades after ~8h: fut.gg returns 200s with
   `liveAuctions: []` for ~87% of requests (normal baseline ~15%). Downstream
   visible_lph=0 → scorer_v3 returns None → every PlayerScore row takes the
   else branch with sell_price=0 and weighted_score=NULL. Verified by a fresh
   Playwright reproducer hitting 14% empty; running scanner hits 87%.

   Fix in playwright_client.py: recycle browser_context every 30 min (age
   trigger) or after 20 consecutive currentPrice>0 AND liveAuctions==[]
   responses (health trigger). Serialized via asyncio.Lock; context build
   extracted into _build_context() so start() and _recycle_context() share
   it.

2. scanner.py:346-348 — v3_result fields came back as Decimal in some paths,
   causing `expected_profit * sph` at line 363 to raise
   "unsupported operand type(s) for *: 'decimal.Decimal' and 'float'". Coerce
   net_profit→int, op_rate→float, expected_profit→float before PlayerScore
   construction.

Verification post-restart: listing_count=0 rate dropped 87% → 16.6% (118/711).
583/711 new player_scores rows have non-NULL weighted_score, sell_price>0,
and scorer_version='v3'. Efficiency values now match the post-fix formula
magnitude (0.0006-0.05) rather than the pre-fix inflated shape.

Root-cause hypothesis for (1) is circumstantial — a fresh Playwright
always worked locally; only difference vs running scanner was session age.
Periodic recycle addresses the observed behavior; if the real cause is
something else (IP fingerprinting, per-host cookies), recycle helps but
won't fully fix it.

+++ src/server/scanner.py
@@ -343,9 +343,9 @@ class ScannerService:
                 #   efficiency                = expected_profit / buy_price
                 #   expected_profit_per_hour  = expected_profit × sales_per_hour
                 #   weighted_score            = scorer_v3 composite ranking score (sell_ratio × sph × net_profit)
-                net_profit = v3_result["net_profit"]
-                op_rate = v3_result["op_sell_rate"]
-                expected_profit = net_profit * op_rate
+                net_profit = int(v3_result["net_profit"])
+                op_rate = float(v3_result["op_sell_rate"])
+                expected_profit = float(net_profit * op_rate)
                 ps = PlayerScore(
                     ea_id=ea_id,
                     scored_at=now,
